Keep checkpoint restore internals in CRIU worker
